requirements.txt
fastapi 
uvicorn 
numpy==1.23.5
paddleocr==2.7.0.2
paddlepaddle==2.5.1
scikit-image
tqdm 
ultralytics
opencv-contrib-python==4.2.0.32
opencv-python==4.6.0.66
opencv-python-headless==4.2.0.34
python-multipart
pymupdf==1.19.0
img2pdf
streamlit
pytesseract
paddleocr==2.7.0.2
paddlepaddle==2.5.1
scikit-image
pymupdf==1.20.2
absl-py==1.4.0
addict==2.4.0
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
attrdict==2.0.1
attrs==21.4.0
Babel==2.12.1
basicsr==1.4.2
bce-python-sdk==0.8.90
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
commonmark==0.9.1
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.2
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
facexlib==0.3.0
faiss-cpu==1.7.1.post2
fastapi==0.103.1
filelock==3.12.3
filterpy==1.4.5
fire==0.5.0
Flask==2.3.3
flask-babel==3.1.0
fonttools==4.42.1
fsspec==2023.9.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.34
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.5
lazy_loader==0.3
lit==16.0.6
llvmlite==0.40.1
lmdb==1.4.1
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
pdf2docx==0.5.6
pdf2image==1.16.3
Pillow==9.5.0
platformdirs==3.10.0
premailer==3.10.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.2.0
rarfile==4.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.1.0
rsa==4.9
safetensors==0.3.3
scipy
transformers
timm
